Feat : Added the correct impulse calculation

+++ TEST2/Vector3D.cpp
@@ -79,7 +79,7 @@ void Vector3D::operator/=(double scalaire) {
 	std::cout << "ERR : Division par 0 impossible" << std::endl;
 }
 
-double Vector3D::operator*(const Vector3D& other) const
+float Vector3D::operator*(const Vector3D& other) const
 {
 	return x * other.x + y * other.y + z * other.z;
 }
